requirements.txt
langchain-openai
langchain-core
langchain-community
requests
beautifulsoup4
faiss-cpu  # or faiss-gpu if you're using a GPU
python-dotenv
